feat(driver): 接入 gateway 包并补强编译诊断与运行时链路
本次提交基于当前工作区全部未提交内容统一整理提交，覆盖构建配置、驱动运行时、编译器诊断、网络网关、新增工具包、测试补强与评审文档落盘。

一、构建与包系统
- 新增 PACKAGE_GATEWAY、PACKAGE_JSON_EXTENSION、PACKAGE_MYUTIL 开关，并补充 seed_random 包接入。
- 调整 build_packages_genfiles.sh 调用方式，显式通过 BASH_EXECUTABLE 执行脚本，修正 PACKAGES_SPEC_FULL 清理变量名错误。
- 在 config.h 中写入实际编译器标志，修正 CXXFLAGS 与版本信息输出。
- 新增 gateway 运行时配置项：gateway port、gateway external、gateway debug、gateway packet size。
- 更新 applies 表，加入 GATEWAY_LOGON 与 GATEWAY_CONNECT apply，重新生成 applies_table.autogen.*。

二、Gateway、JSON 与实用包扩展
- 新增 gateway 包，接入监听、主连接管理、会话创建/销毁、心跳检测、发送接口、interactive 映射联动等能力。
- comm.cc 增加 PORT_TYPE_GATEWAY 发送分支，并在 remove_interactive / exec 过程中补充 gateway 会话清理与映射更新。
- interactive_t 新增 gateway_session_id、gateway_real_ip、gateway_real_port、gateway_master_fd 等字段，并增加 GATEWAY_SESSION / GATEWAY_MSGPACK 标志位。
- 新增 json_extension 包，提供 json_encode/json_decode/read_json/write_json 以及 mapping 排序辅助能力，并为 gateway 输出封包提供 JSON 构建接口。
- 新增 myutil 包，提供 uuid 与字符串缓冲区工具；新增 seed_random 包，提供基于 xoroshiro128++ 的种子随机数接口。
- mainlib 初始化流程补充 gateway/myutil 初始化入口。

三、驱动运行时与平台兼容改进
- init_main 支持 require_backend=false 的离线初始化模式，LPCC / symbol 工具启动时不再强依赖事件后端。
- backend 初始化增加失败诊断日志，DNS 初始化在无 event base 时安全降级。
- Windows 控制台切换为 UTF-8 输入输出，debug_message 在控制台优先走 UTF-16 写出，减少乱码。
- websocket 关闭路径改为更稳妥的 interactive / evbuffer 清理顺序。
- telnet_ext、external、socket_efuns 等处修正 Windows/格式化/分支穿透细节问题。
- add_action 在 f_command 中补充 update_load_av 统计调用。

四、编译器诊断与语义分析补强
- 编译器新增 compile_diagnostic_snapshot 缓存，支持记录 object/file/line/column/message/source/caret 等结构化诊断信息。
- 语法与语义错误改为尽量对齐具体符号位置，新增 diagnostic anchor / override 机制，用于局部变量声明、函数声明、参数类型校验等错误定位。
- mudlib_error_handler 现在会把 compile_error_* 字段注入错误 mapping，便于 mudlib 获取更完整的编译失败上下文。
- 改进 include/header 语法错误、直接语法错误、efun/函数参数类型错误的定位质量。
- 修复 scratchpad 大块分配/重分配实现细节，移除依赖 block[2] 头布局的旧写法。
- 调整 locals 扩容逻辑，同时扩展 locals 与 type_of_locals 两组存储，避免大小不同步。
- 引入 error placeholder 节点判定，避免错误节点继续参与 pop_value 链路。

五、核心运行时逻辑修正
- call_out handle 分配改为显式寻找可用句柄，移除依赖递增 unique 与旧句柄比较的逻辑，避免长生命周期 call_out 句柄失效。
- heartbeat 改为 heartbeats + heartbeats_pending + interval map 组合管理，补齐当前执行对象、动态增删与去重检查逻辑。
- eval_limit 在非 Linux 平台增加基于 steady_clock 的 deadline 回退实现，并补充 check_eval 接口。
- inherit_list / deep_inherit_list 由固定 256 项缓存改为 vector，修复大量继承链场景下的截断问题。
- ed 缩进栈改为静态存储区，避免函数内局部大数组反复构造。
- maximum local variables 运行时配置下限从 64 放宽到 1。

六、协议与输入链路调整
- 为 MUD 包长度解析增加 decode_mud_packet_size 校验函数，统一处理网络字节序与非法长度情况。
- interactive 输入缓冲上限由 1 MiB 调整为 64 KiB，并在读包阶段提前拒绝非法或越界长度。
- Gateway 用户输出改为通过 gateway_send_to_session 回写到网关链路。

七、测试与回归补强
- src/tests/test_lpc.cc 新增编译诊断、include/header 定位、参数类型高亮、heartbeat 生命周期、locals 扩容、scratchpad 拼接、继承列表上限、UTF-8 截断、eval deadline 等多组单元测试。
- 新增 testsuite/include/line_error_bad_header.h、testsuite/single/tests/compiler/fail/line_error_include.c、line_error_direct.c 等编译失败夹具。
- testsuite/command/more.c 调整背景色宏命名，避免缩写冲突。

八、文档与仓库整理
- 新增 docs/driver/uncommitted_review_2026_03_24.md，记录本轮未提交代码评审发现与后续关注项。
- 更新 docs/.vitepress/sidebar.ts，为专项评审记录补充入口。
- 删除 QWEN.md 与 TODO.md。

九、核验记录
- 使用 MSYS2 / MINGW64 重新全新配置构建目录（PACKAGE_DB=OFF）。
- 成功构建目标：lpc_tests。
- 成功执行测试：src/tests/lpc_tests.exe，17/17 通过。

+++ src/packages/external/external.cc
@@ -298,15 +298,17 @@ int external_start(int which, svalue_t *args, svalue_t *arg1, svalue_t *arg2, sv
     error("CreateProcess() in external_start() failed: %s\n", strerror(errno));
     return EESOCKET;
   }
-  debug(external_start, "external_start: Launching external command '%s', pid: %d.\n", cmdline.c_str(),
-                processInfo.dwProcessId);
+  debug(external_start, "external_start: Launching external command '%s', pid: %lu.\n",
+        cmdline.c_str(), static_cast<unsigned long>(processInfo.dwProcessId));
 
   std::thread([=]() {
     WaitForSingleObject(processInfo.hProcess, INFINITE);
     DWORD exitCode = -1;
     // Get the exit code.
     GetExitCodeProcess(processInfo.hProcess, &exitCode);
-    debug(external_start, "external_start: pid: %d exited with %d.\n", processInfo.dwProcessId, exitCode);
+    debug(external_start, "external_start: pid: %lu exited with %lu.\n",
+          static_cast<unsigned long>(processInfo.dwProcessId),
+          static_cast<unsigned long>(exitCode));
     CloseHandle(processInfo.hProcess);
     CloseHandle(processInfo.hThread);
     evutil_closesocket(sv[0]);
